refactor: extract nesting-aware tag parsers into non_nestable_tag.hh
Move try_parse_non_nestable_equal_sign_tag, try_parse_external_tag and
try_parse_link_tag (plus their result types) out of try_parse.hh into a new
header that includes frame_context.hh, so try_parse.hh no longer needs the
ParserFrameContext forward declaration.
